app/requirements/dev.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Pydantic==0.12.0
psycopg2==2.9.9
pyjwt==2.8.0
